requirements.txt
Django==4.2.6
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
django-storages[s3]==1.14.2
asgiref==3.7.2
djoser==2.2.0
python-dotenv==1.0.0
django_otp
django-cors-headers
pyotp==2.9.0
djangorestframework_simplejwt==5.3.0
mysqlclient==2.2.0
django-cors-headers==4.3.0
gunicorn==21.2.0
boto3==1.28.63
django-storages==1.14.2